SessionCipher added

+++ Axolotl/Protocol/PreKeyWhisperMessage.cs
@@ -7,26 +7,34 @@ using Functional.Maybe;
 
 namespace Axolotl.Protocol
 {
-	// Full Complete 
+	// Full Complete
 
 	public class PreKeyWhisperMessage : CiphertextMessage
 	{
-		public int               MessageVersion { get; private set; }
+		public UInt32            MessageVersion { get; private set; }
+
 		public UInt32            RegistrationId { get; private set; }
-		public Maybe<UInt32>	  PreKeyId { get; private set; }
+
+		public Maybe<UInt32>	 PreKeyId { get; private set; }
+
 		public UInt32            SignedPreKeyId { get; private set; }
+
 		public ECPublicKey       BaseKey { get; private set; }
+
 		public IdentityKey       IdentityKey { get; private set; }
+
 		public WhisperMessage    Message { get; private set; }
 
-		private byte[]            _serialized;
+		private byte[] _serialized;
 
 		public PreKeyWhisperMessage(byte[] serialized)
 		{
-			try {
-				MessageVersion = ByteUtil.HighBitsToInt(serialized[0]);
+			try
+			{
+				MessageVersion = (UInt32)ByteUtil.HighBitsToInt(serialized[0]);
 
-				if (MessageVersion > CiphertextMessage.CURRENT_VERSION) {
+				if(MessageVersion > CiphertextMessage.CURRENT_VERSION)
+				{
 					throw new InvalidVersionException("Unknown version: " + MessageVersion);
 				}
 
@@ -38,52 +46,56 @@ namespace Axolotl.Protocol
 					preKeyWhisperMessage = Serializer.Deserialize<WhisperProtos.PreKeyWhisperMessage>(stream);
 				}
 			
-				if ((MessageVersion == 2 && !preKeyWhisperMessage.preKeyId.HasValue)        ||
-				    (MessageVersion == 3 && !preKeyWhisperMessage.signedPreKeyId.HasValue)  ||
-				    preKeyWhisperMessage.baseKey == null                              ||
-				    preKeyWhisperMessage.identityKey == null                       	  ||
-				    preKeyWhisperMessage.message == null)
+				if((MessageVersion == 2 && !preKeyWhisperMessage.preKeyId.HasValue) ||
+				   (MessageVersion == 3 && !preKeyWhisperMessage.signedPreKeyId.HasValue) ||
+				   preKeyWhisperMessage.baseKey == null ||
+				   preKeyWhisperMessage.identityKey == null ||
+				   preKeyWhisperMessage.message == null)
 				{
 					throw new InvalidMessageException("Incomplete message.");
 				}
 
-				_serialized     = serialized;
+				_serialized = serialized;
 				RegistrationId = preKeyWhisperMessage.registrationId.Value;
-				PreKeyId       = preKeyWhisperMessage.preKeyId.Value.ToMaybe();
+				PreKeyId = preKeyWhisperMessage.preKeyId.Value.ToMaybe();
 				SignedPreKeyId = preKeyWhisperMessage.signedPreKeyId.Value; //() ? preKeyWhisperMessage.getSignedPreKeyId() : -1;
-				BaseKey        = Curve.DecodePoint(preKeyWhisperMessage.baseKey, 0);
-				IdentityKey    = new IdentityKey(Curve.DecodePoint(preKeyWhisperMessage.identityKey, 0));
-				Message        = new WhisperMessage(preKeyWhisperMessage.message);
-			} catch (InvalidKeyException e) {
-				throw new InvalidMessageException (e);
-			} catch (LegacyMessageException e) {
-				throw new InvalidMessageException (e);
+				BaseKey = Curve.DecodePoint(preKeyWhisperMessage.baseKey, 0);
+				IdentityKey = new IdentityKey(Curve.DecodePoint(preKeyWhisperMessage.identityKey, 0));
+				Message = new WhisperMessage(preKeyWhisperMessage.message);
+			}
+			catch(InvalidKeyException e)
+			{
+				throw new InvalidMessageException(e);
+			}
+			catch(LegacyMessageException e)
+			{
+				throw new InvalidMessageException(e);
 			}
 		}
 
-		public PreKeyWhisperMessage(int messageVersion, UInt32 registrationId, Maybe<UInt32> preKeyId,
+		public PreKeyWhisperMessage(UInt32 messageVersion, UInt32 registrationId, Maybe<UInt32> preKeyId,
 		                            UInt32 signedPreKeyId, ECPublicKey baseKey, IdentityKey identityKey,
 		                            WhisperMessage message)
 		{
-			MessageVersion        = messageVersion;
+			MessageVersion = messageVersion;
 			RegistrationId = registrationId;
-			PreKeyId       = preKeyId;
+			PreKeyId = preKeyId;
 			SignedPreKeyId = signedPreKeyId;
-			BaseKey        = baseKey;
-			IdentityKey    = identityKey;
-			Message        = message;
+			BaseKey = baseKey;
+			IdentityKey = identityKey;
+			Message = message;
 
 			var preKeyMessage = new WhisperProtos.PreKeyWhisperMessage {
-					signedPreKeyId = SignedPreKeyId,
-					baseKey = BaseKey.Serialize(),
-					identityKey = IdentityKey.Serialize(),
-					message = Message.Serialize(),
-					registrationId = registrationId
+				signedPreKeyId = SignedPreKeyId,
+				baseKey = BaseKey.Serialize(),
+				identityKey = IdentityKey.Serialize(),
+				message = Message.Serialize(),
+				registrationId = registrationId
 			};
 
-			preKeyId.Do (pKid => preKeyMessage.preKeyId = pKid);
+			preKeyId.Do(pKid => preKeyMessage.preKeyId = pKid);
 
-			byte[] versionBytes = { ByteUtil.IntsToByteHighAndLow(MessageVersion, CURRENT_VERSION) };
+			byte[] versionBytes = { ByteUtil.IntsToByteHighAndLow((int)MessageVersion, CURRENT_VERSION) };
 
 			byte[] messageBytes;
 			using(var stream = new MemoryStream())
@@ -95,12 +107,12 @@ namespace Axolotl.Protocol
 			_serialized = ByteUtil.Combine(versionBytes, messageBytes);
 		}
 
-		public override int GetKeyType ()
+		public override int GetKeyType()
 		{
 			return CiphertextMessage.PREKEY_TYPE;
 		}
 
-		public override byte[] Serialize ()
+		public override byte[] Serialize()
 		{
 			return _serialized;
 		}
